Add Test: TestLargeList, TestMassiveSortedList
Added a test that does the same as the original test, only with a larger
list (200 items instead of 50).
Added a second test that stress-tests StringSortedList's performance
(5000 items).
Moved the logic for generating/retrieving test data into its own private
function.
Adjusted the test to account for strings being generated with random
casing.